# service/apigateway/docs/swagger.yaml
definitions:
  errors.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      retryable:
        type: boolean
      status:
        type: string
      trace_id:
        type: string
      type:
        $ref: '#/definitions/errors.ErrorType'
      validations:
        items:
          $ref: '#/definitions/errors.ValidationError'
        type: array
    type: object
  errors.ErrorType:
    enum:
    - INTERNAL
    - NOT_FOUND
    - BAD_REQUEST
    - CONFLICT
    - UNAUTHORIZED
    - FORBIDDEN
    - TIMEOUT
    - UNAVAILABLE
    type: string
    x-enum-varnames:
    - ErrorTypeInternal
    - ErrorTypeNotFound
    - ErrorTypeBadRequest
    - ErrorTypeConflict
    - ErrorTypeUnauthorized
    - ErrorTypeForbidden
    - ErrorTypeTimeout
    - ErrorTypeUnavailable
  errors.ValidationError:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
...
  response.ApiResponsePaginationReview:
    properties:
      data:
        items:
          $ref: '#/definitions/response.ReviewResponse'
        type: array
      message:
        type: string
      pagination:
        $ref: '#/definitions/response.PaginationMeta'
      status:
        type: string
    type: object
  response.ApiResponsePaginationReviewDeleteAt:
    properties:
      data:
        items:
          $ref: '#/definitions/response.ReviewResponseDeleteAt'
        type: array
      message:
        type: string
      pagination:
        $ref: '#/definitions/response.PaginationMeta'
      status:
        type: string
    type: object
...
  response.ApiResponsePaginationReviewsDetail:
    properties:
      data:
        items:
          $ref: '#/definitions/response.ReviewsDetailResponse'
        type: array
      message:
        type: string
      pagination:
        $ref: '#/definitions/response.PaginationMeta'
      status:
        type: string
    type: object
...
  response.PaginationMeta:
    properties:
      current_page:
        type: integer
      page_size:
        type: integer
      total_pages:
        type: integer
      total_records:
        type: integer
    type: object
...
  response.ReviewDetailResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      string:
        type: string
      type:
        type: string
      updated_at:
        type: string
      url:
        type: string
    type: object
...
  response.ReviewResponse:
    properties:
      comment:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      product_id:
        type: integer
      rating:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  response.ReviewResponseDeleteAt:
    properties:
      comment:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      name:
        type: string
      product_id:
        type: integer
      rating:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  response.ReviewsDetailResponse:
    properties:
      comment:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      name:
        type: string
      product_id:
        type: integer
      rating:
        type: integer
      review_detail:
        $ref: '#/definitions/response.ReviewDetailResponse'
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /api/review-query:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all reviews
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of reviews
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationReview'
        "500":
          description: Failed to retrieve review data
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: Find all reviews
      tags:
      - Review Query
  /api/review-query/active:
    get:
      consumes:
      - application/json
      description: Retrieve a list of active reviews
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of active reviews
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationReviewDeleteAt'
        "500":
          description: Failed to retrieve review data
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: Retrieve active reviews
      tags:
      - Review Query
  /api/review-query/merchant/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve reviews for a specific merchant including review details
      parameters:
      - description: Merchant ID
        in: path
        name: id
        required: true
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of reviews with details
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationReviewsDetail'
        "400":
          description: Invalid Merchant ID
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Failed to retrieve review data
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: Find reviews by Merchant ID
      tags:
      - Review Query
  /api/review-query/product/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve reviews for a specific product including review details
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of reviews with details
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationReviewsDetail'
        "400":
          description: Invalid Product ID
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Failed to retrieve review data
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: Find reviews by Product ID
      tags:
      - Review Query
  /api/review-query/trashed:
    get:
      consumes:
      - application/json
      description: Retrieve a list of trashed review records
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of trashed reviews
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationReviewDeleteAt'
        "500":
          description: Failed to retrieve review data
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: Retrieve trashed reviews
      tags:
      - Review Query
...
swagger: "2.0"